Several changes:
1. Minor changes for the single chain factor method (moved factor
creation to factory)

2. Removed some warnings

3. Moved Collapsed Gibbs test to experimental package

4. Updated tutorial
