+++ FigaroLaTeX/Tutorial/FigaroTutorial.tex
@@ -83,7 +83,8 @@
 \include{Sections/9HierReasoning} 
 \include{Sections/10Elements} 
 \include{Sections/11CreateNew} 
-\include{Sections/12Conclusion}
+\include{Sections/12Experimental}
+\include{Sections/13Conclusion}
 
 %----------------------------------------------------------------------------------------
