+++ FigaroLaTeX/Tutorial/Sections/11CreateNew.tex
@@ -55,7 +55,7 @@ A useful operation is to expand all chains in a universe to obtain the complete 
 As usual, the \texttt{universe} argument can be omitted, using the current default universe. Support is provided for algorithms that are based on factors. Variable elimination is one example, but there are many other such algorithms. To create all the factors for an element, use:
 
 \begin{flushleft}
-\texttt{Factory.make(element)}
+\texttt{Factory.makeFactorsForElement(element)}
 \end{flushleft}
 
 The standard procedure to turn a universe into a list of factors is to:
